@@ -1,6 +1,6 @@
 {
   "name": "chris_store_ui",
-  "version": "0.1.1",
+  "version": "0.2.0-rc.1",
   "private": true,
   "scripts": {
     "start": "react-scripts start",
@@ -16,27 +16,30 @@
   },
   "lint-staged": {
     "*.js": [
-      "eslint --fix",
-      "git add",
-      "yarn run test:staged"
+      "eslint --fix"
     ]
   },
   "dependencies": {
     "@fnndsc/chrisstoreapi": "^2.0.2",
+    "@patternfly/patternfly": "^4.59.1",
+    "@patternfly/react-core": "^4.75.2",
+    "@patternfly/react-icons": "^4.7.16",
     "classnames": "^2.2.6",
     "core-js": "^2.5.7",
+    "dompurify": "^2.2.7",
     "email-validator": "^2.0.4",
     "lodash": "^4.17.19",
+    "marked": "^2.0.3",
     "moment": "^2.22.2",
     "patternfly": "^3.58.0",
     "patternfly-react": "^2.24.0",
     "prop-types": "^15.6.2",
-    "react": "16.4.1",
+    "react": "^16.8.0",
     "react-bootstrap": "^0.32.1",
     "react-clipboard.js": "^2.0.0",
     "react-copy-to-clipboard": "^5.0.2",
-    "react-dom": "16.4.2",
-    "react-router-dom": "^4.3.1",
+    "react-dom": "^16.8.0",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.3",
     "rxjs": "^6.3.3",
     "sirv-cli": "^1.0.8",
