@@ -1,9 +1,9 @@
-import React, { Component } from 'react';
-import isEmpty from 'lodash/isEmpty';
-import orderBy from 'lodash/orderBy';
-import PropTypes from 'prop-types';
-import * as sort from 'sortabular';
-import * as resolve from 'table-resolver';
+import React, { Component } from "react";
+import isEmpty from "lodash/isEmpty";
+import orderBy from "lodash/orderBy";
+import PropTypes from "prop-types";
+import * as sort from "sortabular";
+import * as resolve from "table-resolver";
 import {
   actionHeaderCellFormatter,
   customHeaderFormattersDefinition,
@@ -14,17 +14,11 @@ import {
   TABLE_SORT_DIRECTION,
   MenuItem,
   Col,
-  CardTitle,
-  CardBody,
-  Card,
-  CardHeading,
-  Icon,
-  CardLink,
-  CardFooter,
-} from 'patternfly-react';
-
-import BrainyTeammatesPointer from '../../../../assets/img/brainy_teammates-pointer.png';
-import './DashTeamView.css';
+  Icon
+} from "patternfly-react";
+import { CardTitle, CardBody, Card, CardFooter, Grid, GridItem, CardActions } from "@patternfly/react-core";
+import BrainyTeammatesPointer from "../../../../assets/img/brainy_teammates-pointer.png";
+import "./DashTeamView.css";
 
 const DashTeamEmptyState = () => (
   <div className="card-body-content-parent">
@@ -32,15 +26,19 @@ const DashTeamEmptyState = () => (
       <span className="pficon pficon-info" id="no-plugin-info-icon" />
     </div>
     <div className="dash-text-div">
-      <span className="github-plugin-noplugin-title github-team-noplugin-title">Teammates Panel</span>
-      <p className="github-plugin-noplugin-text">In this area, you will be able to add and
-        manage teammates to help you with each plugin.
+      <span className="github-plugin-noplugin-title github-team-noplugin-title">
+        Teammates Panel
+      </span>
+      <p className="github-plugin-noplugin-text">
+        In this area, you will be able to add and manage teammates to help you
+        with each plugin.
       </p>
     </div>
     <div className="card-body-content-child-right">
       <img src={BrainyTeammatesPointer} alt="Click Add Plugin" />
     </div>
-  </div>);
+  </div>
+);
 
 class DashTeamView extends Component {
   constructor(props) {
@@ -51,23 +49,23 @@ class DashTeamView extends Component {
 
     const sortableTransform = sort.sort({
       getSortingColumns,
-      onSort: (selectedColumn) => {
+      onSort: selectedColumn => {
         this.setState({
           sortingColumns: sort.byColumn({
             sortingColumns: this.state.sortingColumns,
             sortingOrder: defaultSortingOrder,
-            selectedColumn,
-          }),
+            selectedColumn
+          })
         });
       },
       // Use property or index dependening on the sortingColumns structure specified
-      strategy: sort.strategies.byProperty,
+      strategy: sort.strategies.byProperty
     });
 
     const sortingFormatter = sort.header({
       sortableTransform,
       getSortingColumns,
-      strategy: sort.strategies.byProperty,
+      strategy: sort.strategies.byProperty
     });
 
     // enables our custom header formatters extensions to reactabular
@@ -78,87 +76,87 @@ class DashTeamView extends Component {
       sortingColumns: {
         name: {
           direction: TABLE_SORT_DIRECTION.ASC,
-          position: 0,
-        },
+          position: 0
+        }
       },
       columns: [
         {
-          property: 'name',
+          property: "name",
           header: {
-            label: 'Name',
+            label: "Name",
             props: {
               index: 0,
               rowSpan: 1,
               colSpan: 1,
-              sort: true,
+              sort: true
             },
             transforms: [sortableTransform],
             formatters: [sortingFormatter],
-            customFormatters: [sortableHeaderCellFormatter],
+            customFormatters: [sortableHeaderCellFormatter]
           },
           cell: {
             props: {
-              index: 0,
+              index: 0
             },
-            formatters: [tableCellFormatter],
-          },
+            formatters: [tableCellFormatter]
+          }
         },
         {
-          property: 'title',
+          property: "title",
           header: {
-            label: 'Position Title',
+            label: "Position Title",
             props: {
               index: 1,
               rowSpan: 1,
               colSpan: 1,
-              sort: true,
+              sort: true
             },
             transforms: [sortableTransform],
             formatters: [sortingFormatter],
-            customFormatters: [sortableHeaderCellFormatter],
+            customFormatters: [sortableHeaderCellFormatter]
           },
           cell: {
             props: {
-              index: 1,
+              index: 1
             },
-            formatters: [tableCellFormatter],
-          },
+            formatters: [tableCellFormatter]
+          }
         },
         {
-          property: 'date_joined',
+          property: "date_joined",
           header: {
-            label: 'Date Joined',
+            label: "Date Joined",
             props: {
               index: 2,
               rowSpan: 1,
               colSpan: 1,
-              sort: true,
+              sort: true
             },
             transforms: [sortableTransform],
             formatters: [sortingFormatter],
-            customFormatters: [sortableHeaderCellFormatter],
+            customFormatters: [sortableHeaderCellFormatter]
           },
           cell: {
             props: {
-              index: 2,
+              index: 2
             },
-            formatters: [tableCellFormatter],
-          },
+            formatters: [tableCellFormatter]
+          }
         },
         {
-          property: 'actions',
+          property: "actions",
           header: {
-            label: 'Actions',
+            label: "Actions",
             props: {
               index: 3,
               rowSpan: 1,
-              colSpan: 2,
+              colSpan: 2
             },
-            formatters: [actionHeaderCellFormatter],
+            formatters: [actionHeaderCellFormatter]
           },
           cell: {
             props: {
-              index: 3,
+              index: 3
             },
             formatters: [
               (value, { rowData }) => [
@@ -166,7 +164,7 @@ class DashTeamView extends Component {
                   <Table.Button
                     onClick={() => console.log(`clicked ${rowData.name}`)}
                   >
-                      Edit
+                    Edit
                   </Table.Button>
                 </Table.Actions>,
                 <Table.Actions key="1">
@@ -177,13 +175,13 @@ class DashTeamView extends Component {
                     <MenuItem divider />
                     <MenuItem>Separated link</MenuItem>
                   </Table.DropdownKebab>
-                </Table.Actions>,
-              ],
-            ],
-          },
-        },
+                </Table.Actions>
+              ]
+            ]
+          }
+        }
       ],
-      rows: [{}],
+      rows: [{}]
     };
   }
 
@@ -196,21 +194,18 @@ class DashTeamView extends Component {
       columns,
       sortingColumns,
       sort: orderBy,
-      strategy: sort.strategies.byProperty,
+      strategy: sort.strategies.byProperty
     })(rows);
 
     return (
-      <Col sm={12}>
+      <Grid sm={12}>
+        <GridItem>
         <Card>
-          <CardHeading>
-            <CardTitle>
-              Teammates
-            </CardTitle>
-          </CardHeading>
+          <CardTitle>Teammates</CardTitle>
           <CardBody>
-            {showEmptyState ?
+            {showEmptyState ? (
               <DashTeamEmptyState />
-              :
+            ) : (
               <React.Fragment>
                 <Table.PfProvider
                   hover
@@ -219,46 +214,49 @@ class DashTeamView extends Component {
                   components={{
                     header: {
                       cell: cellProps =>
-                        (this.customHeaderFormatters({
+                        this.customHeaderFormatters({
                           cellProps,
                           columns,
-                          sortingColumns,
-                        })),
-                    },
-                  }
-                  }
+                          sortingColumns
+                        })
+                    }
+                  }}
                 >
                   <Table.Header headerRows={resolve.headerRows({ columns })} />
                   <Table.Body
                     rows={sortedRows}
                     rowKey="id"
                     onRow={() => ({
-                        role: 'row',
-                      }
-                    )}
+                      role: "row"
+                    })}
                   />
                 </Table.PfProvider>
               </React.Fragment>
-            }
+            )}
           </CardBody>
-          { !showEmptyState &&
-          <CardFooter className="card-footer">
-            <CardLink href="#" icon={<Icon type="pf" name="add-circle-o" />}>
-              Add Teammate
-            </CardLink>
-          </CardFooter>
-          }
+          {!showEmptyState && (
+            <CardFooter
+              className="card-footer"
+            >
+              <CardActions>
+                <Icon type="pf" name="add-circle-o" />
+                Add Teammate
+              </CardActions>
+              
+            </CardFooter>
+          )}
         </Card>
-      </Col>
+        </GridItem>
+      </Grid>
     );
   }
 }
 DashTeamView.propTypes = {
-  plugins: PropTypes.arrayOf(PropTypes.object),
+  plugins: PropTypes.arrayOf(PropTypes.object)
 };
 
 DashTeamView.defaultProps = {
-  plugins: [],
+  plugins: []
 };
 
 export default DashTeamView;
