@@ -3,20 +3,17 @@ import { Redirect } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import {
   Form,
-  FormGroup,
-  ControlLabel,
-  FormControl,
-  HelpBlock,
-  Button,
   Spinner,
-} from 'patternfly-react';
+} from '@patternfly/react-core';
+import Button from '../../../Button';
 import _ from 'lodash';
 import StoreClient from '@fnndsc/chrisstoreapi';
 import { validate } from 'email-validator';
 import './DeveloperSignup.css';
 import ChrisStore from '../../../../store/ChrisStore';
+import FormInput from '../../../FormInput';
 
-/* inspired by http://bit.ly/2KycT4G */
+/* inspired by https://github.com/Modernizr/Modernizr/blob/v3/feature-detects/touchevents.js */
 const isTouchDevice = () => {
   if (('ontouchstart') in window || (window.DocumentTouch && document instanceof window.DocumentTouch)) {
     return true;
@@ -42,11 +39,15 @@ export class DeveloperSignup extends Component {
         message: '',
         controls: '',
       },
+      username: '',
+      email:'',
+      password: '',
+      passwordConfirm: '',
     };
   }
 
-  handleChange(event) {
-    this.setState({ [event.target.name]: event.target.value });
+  handleChange(value, name) {
+    this.setState({ [name]: value });
   }
 
   handleSubmit(event) {
@@ -177,86 +178,69 @@ export class DeveloperSignup extends Component {
     return (
       <Form onSubmit={this.handleSubmit} noValidate>
         <p>{loading ? 'Creating' : 'Create'} a ChRIS Developer account:</p>
-        <FormGroup
-          controlId="username"
-          validationState={error.controls.includes('username') ? 'error' : null}
-        >
-          <ControlLabel>
-            Username
-          </ControlLabel>
-          <FormControl
-            type="text"
-            autoComplete="off"
-            autoFocus={!isTouchDevice()}
-            onChange={this.handleChange}
-            name="username"
-            disabled={disableControls}
-          />
-          <HelpBlock>
-            { error.controls.includes('username') ? error.message : 'Enter your username' }
-          </HelpBlock>
-        </FormGroup>
-        <FormGroup
-          controlId="email"
-          validationState={error.controls.includes('email') ? 'error' : null}
-        >
-          <ControlLabel>
-            Email
-          </ControlLabel>
-          <FormControl
-            type="email"
-            autoComplete="off"
-            onChange={this.handleChange}
-            name="email"
-            disabled={disableControls}
-          />
-          <HelpBlock>
-            { error.controls.includes('email') ? error.message : 'Enter your email' }
-          </HelpBlock>
-        </FormGroup>
-        <FormGroup
-          controlId="password"
-          validationState={error.controls.includes('password') ? 'error' : null}
-        >
-          <ControlLabel>
-            Password
-          </ControlLabel>
-          <FormControl
-            type="password"
-            autoComplete="new-password"
-            onChange={this.handleChange}
-            name="password"
-            disabled={disableControls}
-          />
-          <HelpBlock>
-            { error.controls.includes('password') ? error.message : 'Enter your password' }
-          </HelpBlock>
-        </FormGroup>
-        <FormGroup
-          controlId="password-confirm"
-          validationState={error.controls.includes('confirmation') ? 'error' : null}
-        >
-          <ControlLabel>
-            Password Confirmation
-          </ControlLabel>
-          <FormControl
-            type="password"
-            autoComplete="new-password"
-            onChange={this.handleChange}
-            name="passwordConfirm"
-            disabled={disableControls}
-          />
-          <HelpBlock>
-            { error.controls.includes('confirmation') ? error.message : 'Confirm your password' }
-          </HelpBlock>
-        </FormGroup>
-        <Spinner loading={loading} size="md" inline>
-          {
-            <Button bsStyle="primary" bsSize="large" type="submit" disabled={disableControls}>
+        <FormInput
+          formLabel="Username"
+          fieldId="username"
+          validationState={error.controls.includes('username') ? 'error' : 'default'}
+          helperText="Enter your username"
+          inputType="text"
+          id="username"
+          fieldName="username"
+          value={this.state.username}
+          autofocus={!isTouchDevice}
+          onChange={(val) => this.handleChange(val, 'username')}
+          disableControls={disableControls}
+          error={error}
+        />
+        <FormInput
+          formLabel="Email"
+          fieldId="email"
+          validationState={error.controls.includes('email') ? 'error' : 'default'}
+          helperText="Enter you email"
+          inputType="email"
+          id="email"
+          fieldName="email"
+          value={this.state.email}
+          onChange={(val) => this.handleChange(val, 'email')}
+          disableControls={disableControls}
+          error={error}
+        />
+        <FormInput
+          formLabel="Password"
+          fieldId="password"
+          validationState={error.controls.includes('password') ? 'error' : 'default'}
+          helperText="Enter your password"
+          inputType="password"
+          id="password"
+          fieldName="password"
+          value={this.state.password}
+          onChange={(val) => this.handleChange(val, 'password')}
+          disableControls={disableControls}
+          error={error}
+        />
+        <FormInput
+          formLabel="Password Conformation"
+          fieldId="password-confirm"
+          validationState={error.controls.includes('confirmation') ? 'error' : 'default'}
+          helperText="Confirm your password"
+          inputType="password"
+          id="password"
+          fieldName="passwordConfirm"
+          value={this.state.passwordConfirm}
+          onChange={(val) => this.handleChange(val, 'passwordConfirm')}
+          disableControls={disableControls}
+          error={error}
+        />
+        {loading ? <Spinner size="md"/> : (
+          <Button 
+            variant="primary"
+            type="submit" 
+            loading={disableControls}
+          >
             Create Account
-            </Button>
-          }
-        </Spinner>{loading && <span className="developer-signup-creating">  Creating Account</span>}
+          </Button>
+        )}
+        {loading && <span className="developer-signup-creating">Creating Account</span>}
       </Form>);
   }
 }
