@@ -1,202 +1,170 @@
-/* eslint-disable react/jsx-indent-props */
-/* eslint-disable react/jsx-indent */
-import React from 'react';
+import React, { useState, useEffect, useCallback } from 'react';
+import { Grid, GridItem, Tabs, Tab, TabTitleText } from '@patternfly/react-core';
 import PropTypes from 'prop-types';
-import { Link } from 'react-router-dom';
-import {
-  Icon, Grid, Nav, NavItem, TabContainer, TabContent, TabPane,
-} from 'patternfly-react';
-import './PluginBody.css';
-import { CopyURLButton } from '../../../general/CopyURLButton';
+import marked from 'marked';
+import { sanitize } from 'dompurify';
 
-const PluginBody = ({ pluginData }) => (
-  <div className="plugin-body">
-    <div className="container-fluid container-cards-pf">
-      <div className="row row-cards-pf">
-        <Grid>
-          <Grid.Row>
-            <Grid.Col sm={12}>
-              <div className="card-pf">
-                <TabContainer id="basic-tabs" defaultActiveKey={1}>
-                  <div>
-                    <Nav className="nav nav-tabs nav-tabs-pf plugin-body-nav-tabs">
-                      <NavItem eventKey={1}>
-                        Overview
-                      </NavItem>
-                      <NavItem eventKey={2}>
-                        Parameters
-                      </NavItem>
-                      <NavItem eventKey={3}>
-                        Versions
-                      </NavItem>
-                    </Nav>
-                    <TabContent animation className="plugin-tab">
-                      <TabPane eventKey={1}>
-                      <div className="plugin-body-combined-cols">
-                        <div className="plugin-body-main-col">
-                          <div className="plugin-body-readme">
-                            README.rst
-                          </div>
-                          <h1 className="plugin-body-title">
-                            {pluginData.title}
-                          </h1>
-                          <div className="plugin-body-toc">
-                            <h3>Table of Contents</h3>
-                            <ul>
-                              <li>
-                                <Link
-                                  href={pluginData.title}
-                                  to={pluginData.title}
-                                >
-                                  Abstract
-                                </Link>
-                              </li>
-                              <li>
-                                <Link
-                                  href={pluginData.title}
-                                  to={pluginData.title}
-                                >
-                                  Synopsis
-                                </Link>
-                              </li>
-                              <li>
-                                <Link
-                                  href={pluginData.title}
-                                  to={pluginData.title}
-                                >
-                                  Run
-                                </Link>
-                                <ul>
-                                  <li>
-                                    <Link
-                                      href={pluginData.title}
-                                      to={pluginData.title}
-                                    >
-                                      Using PyPi
-                                    </Link>
-                                  </li>
-                                  <li>
-                                    <Link
-                                      href={pluginData.title}
-                                      to={pluginData.title}
-                                    >
-                                      Using
-                                      <code>docker run</code>
-                                    </Link>
-                                  </li>
-                                </ul>
-                              </li>
-                              <li>
-                                <Link
-                                  href={pluginData.title}
-                                  to={pluginData.title}
-                                >
-                                  Examples
-                                </Link>
-                                <ul>
-                                  <li>
-                                    <Link
-                                      href={pluginData.title}
-                                      to={pluginData.title}
-                                    >
-                                      Check available pre-processed runs
-                                    </Link>
-                                  </li>
-                                  <li>
-                                    <Link
-                                      href={pluginData.title}
-                                      to={pluginData.title}
-                                    >
-                                      Copy the default for a selected pre-processed run
-                                    </Link>
-                                  </li>
-                                </ul>
-                              </li>
-                            </ul>
-                          </div>
-                          <div className="plugin-body-content">
-                            <h2>Abstract</h2>
-                            <hr />
-                            <p>
-                              {/* eslint-disable-next-line */}
-                              <code>freesurfer_pp_moc.py</code> is a dummy FreeSurver plugin/container that is prepopulated with the results of several priori FreeSurfer runs. For a given run, this script will simply copy elements of one of these prior runs to the output directory.
-                            </p>
-                            <h2>Synopsis</h2>
-                            <hr />
-                            <p>
-                              <code>
-                                python freesurfer_pp.py
-                                {/* eslint-disable-next-line */}
-                                Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
-                              </code>
-                            </p>
-                          </div>
-                        </div>
-                        <div className="plugin-body-side-col">
+import './PluginBody.css';
+import CopyURLButton  from '../../../general/CopyURLButton';
+import Notification from '../../../Notification';
+import HttpApiCallError from '../../../../errors/HttpApiCallError';
+import { GithubAPIRepoError,GithubAPIProfileError, GithubAPIReadmeError } from '../../../../errors/GithubError';
+
+
+const removeEmail = (author) => author.replace(/( ?<.*>)/g, '');
+
+const PluginBody = ({ pluginData }) => {
+  const [activeTab, setActiveTab] = useState(1);
+  const handleTabClick = (event, tabIndex) => setActiveTab(tabIndex);
+
+  const [repoData, setRepoData] = useState();
+  const [readme, setReadme] = useState();
+  const setReadmeHTML = $ => setReadme(sanitize($))
+
+  const [errors, setErrors] = useState([]);
+  const showNotifications = useCallback((error) => {
+    setErrors([ ...errors, error.message ])
+  }, [errors])
+
+  const fetchReadme = useCallback(async (repo) => {
+    const profile =  await fetch(`https://api.github.com/repos/${repo}/community/profile`)
+    if (!profile.ok) throw new GithubAPIProfileError(repo, profile)
+
+    const url = await fetch((await profile.json()).files.readme.url)
+    if (!url.ok) throw new GithubAPIReadmeError(repo, url)
+
+    const { download_url } = (await url.json())
+    const file = await (await fetch(download_url)).text()
+    const type = download_url.split('.').reverse().shift()
+
+    return { file, type }
+  }, [])
+
+  const fetchRepoData = useCallback(async (repo) => {
+    const data = await fetch(`https://api.github.com/repos/${repo}`)
+    if (!data.ok) throw new GithubAPIRepoError(repo, data)
+
+    return await data.json()
+  }, [])
+
+  useEffect(() => {
+    async function fetchRepo() {
+      try {
+        const data = await fetchRepoData(expectRepoName)
+        setRepoData(data)
+
+        const { file, type } = await fetchReadme(expectRepoName)
+        if (type === 'md' || type === 'rst') 
+          setReadmeHTML(marked(file))
+        else
+          setReadmeHTML(file)
+      } catch (error) {
+        showNotifications(new HttpApiCallError(error))
+      }
+    }
+
+    const expectRepoName = pluginData.public_repo.split('github.com/')[1];
+    if (expectRepoName)
+      fetchRepo()
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [fetchReadme, fetchRepoData, pluginData.public_repo])
+
+  return (
+    <React.Fragment>
+      {
+        errors.map((message, index) => (
+          <Notification
+            title={message}
+            position='top-right'
+            variant='danger'
+            closeable
+            onClose={() => {
+              setErrors(errors.splice(index))
+            }}
+          />
+        ))
+      }
+
+      <div className="plugin-body">
+        <div className="container-fluid container-cards-pf">
+          <div className="row row-cards-pf">
+            <div className="card-pf">
+              <Tabs activeKey={activeTab} onSelect={handleTabClick}>
+                <Tab eventKey={1} title={<TabTitleText>Overview</TabTitleText>}>
+                  <Grid hasGutter className="plugin-body-main">
+                    <GridItem md={8} sm={12}>
+                      <div className="plugin-body-main-col">
+                        <div style={{ color: "gray" }}>README</div>
+                        { readme ? <div dangerouslySetInnerHTML={{ __html: readme }}></div> : null }
+                      </div>
+                    </GridItem>
+                    <GridItem md={4} sm={12}>
+                      <div className="plugin-body-side-col">
                         <div className="plugin-body-copy-url">
-                        <CopyURLButton text={pluginData.url} />
+                          <CopyURLButton className="pf-c-button pf-m-primary" text={pluginData.url} />
                         </div>
-                          <div className="plugin-body-public-repo">
-                            <h4>Public Repo:</h4>
-                            <a href={pluginData.public_repo}>
-                              {pluginData.public_repo}
+                        <div className="plugin-body-detail-section">
+                          <h4>Repository</h4>
+                          <a href={pluginData.public_repo}>
+                            {pluginData.public_repo}
+                          </a>
+                        </div>
+                        <div className="plugin-body-detail-section">
+                          <h4>Contributors</h4>
+                          <a href={pluginData.authorURL}
+                            className="pf-m-link" type="button">
+                            <span className="pf-c-button__icon pf-m-start">
+                              <i className="fas fa-user" aria-hidden="true"></i>
+                            </span>{' '}
+                            {removeEmail(pluginData.authors)}
+                          </a>
+                          <div className="plugin-body-contributors-all">
+                            <br/>
+                            <a className="pf-m-link" href={`${pluginData.public_repo}/graphs/contributors`}>
+                              View all contributors
                             </a>
                           </div>
-                          <div className="plugin-body-documentation">
-                            <h4>Documentation:</h4>
-                            {/* eslint-disable-next-line */}
-                            <a>
-                              readthedocs.com/freesurfer
-                            </a>
-                          </div>
-                          <div className="plugin-body-contributors">
-                            <h4>Contributors (20):</h4>
-                            <Icon name="user" size="lg" />
-                            <Link
-                              href={pluginData.authorURL}
-                              to={pluginData.authors}
-                            >
-                              {pluginData.authors}
-                            </Link>
-                            <div className="plugin-body-contributors-all">
-                              {/* eslint-disable-next-line */}
-                              <a>
-                                View all contributors...
-                              </a>
-                            </div>
-                          </div>
-                          <div className="plugin-body-id">
-                            <h4>Plugin ID:</h4>
-                            {pluginData.id}
-                          </div>
-                          <div className="plugin-body-license">
-                            <h4>License:</h4>
-                            {pluginData.license}
-                          </div>
-                          <div className="plugin-body-content-type">
-                            <h4>Content Type:</h4>
-                            {pluginData.type}
-                          </div>
-                          <div className="plugin-body-date-added">
-                            <h4>Date added:</h4>
-                            7 July 2020
-                          </div>
+                        </div>
+                        <div className="plugin-body-detail-section">
+                          <h4>Plugin ID</h4>
+                          {pluginData.id}
+                        </div>
+                        <div className="plugin-body-detail-section">
+                          <h4>License</h4>
+                          { repoData ? repoData.license.name : pluginData.license }
+                        </div>
+                        <div className="plugin-body-detail-section">
+                          <h4>Content Type</h4>
+                          {pluginData.type}
+                        </div>
+                        <div className="plugin-body-detail-section">
+                          <h4>Date added</h4>
+                          {(new Date(pluginData.creation_date)).toDateString()}
                         </div>
                       </div>
-                      </TabPane>
-                      <TabPane eventKey={2}>Parameters content</TabPane>
-                      <TabPane eventKey={3}>Versions content</TabPane>
-                    </TabContent>
-                  </div>
-                </TabContainer>
-              </div>
-            </Grid.Col>
-          </Grid.Row>
-        </Grid>
+                    </GridItem>
+                  </Grid>
+                </Tab>
+
+                <Tab eventKey={2} title={<TabTitleText>Parameters</TabTitleText>}>
+                  <Grid hasGutter className="plugin-body-main">
+                    <GridItem sm={12}>Parameters content</GridItem>
+                  </Grid>
+                </Tab>
+
+                <Tab eventKey={3} title={<TabTitleText>Versions</TabTitleText>}>
+                  <Grid hasGutter className="plugin-body-main">
+                    <GridItem sm={12}>Versions content</GridItem>
+                  </Grid>
+                </Tab>
+              </Tabs>
+            </div>
+          </div>
+        </div>
       </div>
-    </div>
-  </div>
-);
+    </React.Fragment>
+  );
+}
 
 PluginBody.propTypes = {
   pluginData: PropTypes.shape({
