@@ -1,19 +1,25 @@
-import React, { Component } from "react";
-import PropTypes from "prop-types";
-import Client from "@fnndsc/chrisstoreapi";
-import { DropdownButton, MenuItem } from "react-bootstrap";
-import PluginItem from "./components/PluginItem/PluginItem";
-import LoadingPluginItem from "./components/LoadingPluginItem/LoadingPluginItem";
-import PluginsCategories from "./components/PluginsCategories/PluginsCategories";
-import "./Plugins.css";
-import LoadingContainer from "../LoadingContainer/LoadingContainer";
-import LoadingContent from "../LoadingContainer/components/LoadingContent/LoadingContent";
-import ChrisStore from "../../store/ChrisStore";
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+import Client from '@fnndsc/chrisstoreapi';
+import { DropdownButton, MenuItem } from 'react-bootstrap';
 
-// ==============================
-// ------ PLUGINS COMPONENT -----
-// ==============================
+import PluginItem from './components/PluginItem/PluginItem';
+import LoadingPluginItem from './components/LoadingPluginItem/LoadingPluginItem';
+import PluginsCategories from './components/PluginsCategories/PluginsCategories';
+import LoadingContainer from '../LoadingContainer/LoadingContainer';
+import LoadingContent from '../LoadingContainer/components/LoadingContent/LoadingContent';
+import ChrisStore from '../../store/ChrisStore';
+import HttpApiCallError from '../../errors/HttpApiCallError';
+import Notification from '../Notification';
 
+import './Plugins.css';
+
+const CATEGORIES = ['FreeSurfer', 'MRI', 'Segmentation', 'copy'];
+
+/**
+ * A page showing a list of ChRIS plugins, according to the search
+ * specified in the URI's query string.
+ */
 export class Plugins extends Component {
   constructor(props) {
     super(props);
@@ -22,54 +28,41 @@ export class Plugins extends Component {
     const auth = { token: props.store.get("authToken") };
     this.client = new Client(storeURL, auth);
 
-    this.mounted = false;
+    const categories = new Map();
+    CATEGORIES.forEach((name) => categories.set(name, 0));
+
     this.state = {
-      pluginList: null,
+      errorMsg: null,
       starsByPlugin: {},
-      categories: [
-        { name: "Visualization", length: 0 },
-        { name: "Modeling", length: 0 },
-        { name: "Statistical Operation", length: 0 },
-        { name: "FreeSurfer", length: 0 },
-        { name: "MRI Processing", length: 0 },
-      ],
+      selectedCategory: null,
+      categories: categories
     };
   }
 
-  componentWillMount() {
-    this.mounted = true;
-  }
-
+  /**
+   * Search for plugin data from the backend.
+   */
   async componentDidMount() {
-    try {
-      const plugins = await this.fetchPlugins()
-      /**
-       * Accumulate counts of categories from fetched plugins
-       */
-      let { categories } = this.state;
-      for (const [index, { name, length }] of categories.entries()) {
-        categories[index].length = plugins.reduce(
-          (count, current) => (name === current.category) 
-            ? ++count
-            : count, 
-          length
-        );
-      }
-      
-      this.setState({ categories });
-    } catch (err) {
-      console.error(err);
-    }
+    await this.refreshPluginList();
+  }
 
-    if (this.isLoggedIn()) {
-      this.fetchPluginStars();
+  /**
+   * Re-fetch the list of plugins if the input was changed
+   * in the NarBar's search bar.
+   */
+  async componentDidUpdate(prevProps) {
+    if (this.props.location !== prevProps.location) {
+      await this.refreshPluginList();
     }
   }
 
-  componentWillUnmount() {
-    this.mounted = false;
-  }
-
+  /**
+   * Add a star next to the plugin visually.
+   * (Does not necessarily send to the backend that the plugin is a favorite).
+   *
+   * @param pluginId
+   * @param star
+   */
   setPluginStar(pluginId, star) {
     this.setState({
       starsByPlugin: {
@@ -79,10 +72,33 @@ export class Plugins extends Component {
     });
   }
 
+  /**
+   * Remove a star from next to the plugin visually.
+   * (Does not necessarily send to the backend that the plugin was unfavorited.)
+   * @param pluginId
+   */
   removePluginStar(pluginId) {
     this.setPluginStar(pluginId, undefined);
   }
 
+  /**
+   * Show a notification for some network error.
+   * @param error error message
+   */
+  showNotifications = (error) => {
+    console.error(error);
+    this.setState({
+      errorMsg: error.message,
+    })
+  }
+
+  /**
+   * Mark a plugin as a favorite by showing a star next to it and
+   * notifying the backend of this change.
+   *
+   * @param plugin
+   * @return {Promise<void>}
+   */
   async favPlugin(plugin) {
     // Early state change for instant visual feedback
     this.setPluginStar(plugin.id, {});
@@ -94,10 +110,16 @@ export class Plugins extends Component {
       this.setPluginStar(plugin.id, star.data);
     } catch (err) {
       this.removePluginStar(plugin.id);
-      console.error(err);
+      this.showNotifications(new HttpApiCallError(err));
     }
   }
 
+  /**
+   * Unfavorite a plugin by removing its star and notifying the backend.
+   *
+   * @param plugin
+   * @return {Promise<void>}
+   */
   async unfavPlugin(plugin) {
     const previousStarState = { ...this.state.starsByPlugin[plugin.id] };
 
@@ -109,61 +131,88 @@ export class Plugins extends Component {
       await star.delete();
     } catch (err) {
       this.setPluginStar(plugin.id, previousStarState);
-      console.error(err);
+      this.showNotifications(new HttpApiCallError(err));
     }
   }
 
-  fetchPlugins = () => {
+  /**
+   * 1. Fetch the list of plugins based on the search query.
+   * 2. Accumulate information about categories.
+   * 3. Call setState
+   * 4. If user is logged in, get information about their favorite plugins.
+   */
+  refreshPluginList = async () => {
     const params = new URLSearchParams(window.location.search)
-    const name = params.get('q') //get value searched from the URL
+    const name = params.get('q');
     const searchParams = {
       limit: 20,
       offset: 0,
-      name_title_category:name,
+      name_title_category: name,
     };
 
-    return new Promise(async (resolve, reject) => {
-      let plugins;
-      try {
-        // add plugins to pluginList as they are received
-        plugins = await this.client.getPlugins(searchParams);
-        if (this.mounted) {
-          this.setState((prevState) => {
-            // const prevPluginList = prevState.pluginList
-            //   ? prevState.pluginList
-            //   : [];
-            // const nextPluginList = prevPluginList.concat(plugins.data);
-            return { pluginList: plugins.data };
-          });
-        }
-      } catch (e) {
-        return reject(e);
-      }
-      
-      return resolve(plugins.data);
-    });
-  }
+    let plugins;
 
-  async fetchPluginStars() {
-    const stars = await this.client.getPluginStars();
-
-    const starsByPlugin = {};
-    stars.data.forEach((star) => {
-      const pluginId = star.meta_id;
-      starsByPlugin[pluginId] = star;
-    });
-
-    this.setState({ starsByPlugin });
-  }
-
-  handlePluginFavorited(plugin) {
-    if (this.isLoggedIn()) {
-      return this.isFavorite(plugin)
-        ? this.unfavPlugin(plugin)
-        : this.favPlugin(plugin);
+    try {
+      plugins = await this.client.getPlugins(searchParams);
+    } catch(error) {
+      this.showNotifications(new HttpApiCallError(error));
+      return;
     }
 
-    return Promise.resolve();
+    // reset category counts
+    const categories = this.state.categories;
+    for (const name of categories.keys()) {
+      categories.set(name, 0);
+    }
+
+    // count the frequency of plugins which belong to categories
+    for (const { category } of plugins.data) {
+      // TODO make category counting case insensitive
+      const currentCount = categories.get(category);
+      if (currentCount !== undefined) {
+        categories.set(category, currentCount + 1);
+      }
+    }
+
+    // plugin list and category list are available always, even if not logged in
+    const nextState = {
+      pluginList: plugins.data,
+      categories: categories
+    };
+
+    if (this.isLoggedIn()) {
+      try{
+        const stars = await this.client.getPluginStars();
+        const starsByPlugin = {};
+        stars.data.forEach((star) => {
+          const pluginId = star.meta_id;
+          starsByPlugin[pluginId] = star;
+        });
+        nextState.starsByPlugin = starsByPlugin;
+      } catch(error) {
+        this.showNotifications(new HttpApiCallError(error));
+      }
+    }
+
+    // finally update the state once with pluginList, categories, and maybe starsByPlugin
+    this.setState(nextState);
+  }
+
+  /**
+   * Favorite the plugin if it is not a favorite, or remove from favorites if already a favorite.
+   *
+   * @param plugin
+   * @return {Promise<void>}
+   */
+  togglePluginFavorited(plugin) {
+    if (this.isLoggedIn()) {
+      if (this.isFavorite(plugin)) {
+        this.unfavPlugin(plugin);
+      }
+      else {
+        this.favPlugin(plugin);
+      }
+    }
   }
 
   handleCategorySelect = async (category) => {
@@ -185,8 +234,22 @@ export class Plugins extends Component {
     return this.props.store ? this.props.store.get("isLoggedIn") : false;
   }
 
+  /**
+   * Show only plugins which are part of this category.
+   *
+   * @param name name of category
+   */
+  handleCategorySelect = (name) => {
+    this.setState({selectedCategory: name});
+  }
+
   render() {
-    const { pluginList, categories } = this.state;
+    const { pluginList, categories, selectedCategory } = this.state;
+
+    // convert map into the data structure expected by <PluginsCategories />
+    const categoryEntries = Array.from(categories.entries(), ([name, count]) => ({
+      name: name, length: count
+    }));
 
     // Remove email from author
     const removeEmail = (author) => author.replace(/( ?\(.*\))/g, "");
@@ -196,24 +259,32 @@ export class Plugins extends Component {
 
     // Render the pluginList if the plugins have been fetched
     if (pluginList) {
-      pluginListBody = pluginList.map((plugin) => (
-        <PluginItem
-          title={plugin.title}
-          id={plugin.id}
-          name={plugin.name}
-          author={removeEmail(plugin.authors)}
-          creationDate={plugin.creation_date}
-          key={`${plugin.name}-${plugin.id}`}
-          isLoggedIn={this.isLoggedIn()}
-          isFavorite={this.isFavorite(plugin)}
-          onStarClicked={async () => this.handlePluginFavorited(plugin)}
-        />
+      pluginListBody = pluginList
+        .filter((plugin) => {
+          if (selectedCategory) {
+            return plugin.category === selectedCategory;
+          }
+          return true;
+        })
+        .map((plugin) => (
+          <PluginItem
+            title={plugin.title}
+            id={plugin.id}
+            name={plugin.name}
+            author={removeEmail(plugin.authors)}
+            creationDate={plugin.creation_date}
+            key={`${plugin.name}-${plugin.id}`}
+            isLoggedIn={this.isLoggedIn()}
+            isFavorite={this.isFavorite(plugin)}
+            onStarClicked={() => this.togglePluginFavorited(plugin)}
+          />
       ));
 
       pluginsFound = (
         <span className="plugins-found">{pluginList.length} plugins found</span>
       );
-    } else {
+    }
+    else {
       // Or else show the loading placeholders
       pluginsFound = (
         <LoadingContainer>
@@ -235,6 +306,15 @@ export class Plugins extends Component {
 
     return (
       <div className="plugins-container">
+        {this.state.errorMsg && (
+          <Notification
+            title={this.state.errorMsg}
+            position='top-right'
+            variant='danger'
+            closeable
+            onClose={()=>this.setState({ errorMsg: null })}
+          />
+        )}
         <div className="plugins-stats">
           <div className="row plugins-stats-row">
             {pluginsFound}
@@ -244,8 +324,8 @@ export class Plugins extends Component {
           </div>
         </div>
         <div className="row plugins-row">
-          <PluginsCategories categories={categories} 
-            onSelect={this.handleCategorySelect} 
+          <PluginsCategories categories={categoryEntries}
+            onSelect={this.handleCategorySelect}
           />
           <div className="plugins-list">{pluginListBody}</div>
         </div>
