@@ -12,7 +12,10 @@ const ProtectedRoute = ({ component: Component, ...rest }) => (
       if (isLoggedIn) {
         return <Component {...props} />;
       }
-        return <Redirect to="/signin" />;
+        return <Redirect to={{
+          pathname: "/signin",
+          state: {from: props.location},
+        }} />;
     }}
   />
 );
