@@ -1,14 +1,15 @@
 import React from 'react';
-import { Icon } from 'patternfly-react';
+import { AngleDownIcon } from '@patternfly/react-icons';
 import './WelcomeCTA.css';
 import WelcomeFeature from '../WelcomeFeature/WelcomeFeature';
 
 /* plugin images */
+import infantLogo from '../../../../assets/img/plugins/infant_90.png';
 import freesurferLogo from '../../../../assets/img/plugins/freesurfer_90.png';
-import tensorflowLogo from '../../../../assets/img/plugins/tensorflow_90.png';
-import fslLogo from '../../../../assets/img/plugins/fsl_90.png';
-import rhoanaLogo from '../../../../assets/img/plugins/rhoana_90.png';
-import slicerLogo from '../../../../assets/img/plugins/slicer_90.png';
+import biomediaLogo from '../../../../assets/img/plugins/biomedia_90.png';
+import antLogo from '../../../../assets/img/plugins/ant_90.png';
+import civetLogo from '../../../../assets/img/plugins/civet_90.png';
+
 
 const WelcomeCTA = () => (
   <div className="welcome-cta">
@@ -24,14 +25,14 @@ const WelcomeCTA = () => (
         Apps available in the ChRIS store include:
       </div>
       <div className="row welcome-cta-featured-container">
-        <WelcomeFeature name="FreeSurfer" img={freesurferLogo} />
-        <WelcomeFeature name="TensorFlow" img={tensorflowLogo} />
-        <WelcomeFeature name="FMRIB Software Library" img={fslLogo} />
-        <WelcomeFeature name="RhoANA" img={rhoanaLogo} />
-        <WelcomeFeature name="Slicer" img={slicerLogo} />
+        <WelcomeFeature name="Fastsurfer" img={freesurferLogo} url='https://chrisstore.co/plugin/44'/>
+        <WelcomeFeature name="Infant-FreeSurfer" img={infantLogo} url='https://chrisstore.co/plugin/78'/>
+        <WelcomeFeature name="IRTK reconstruction" img={biomediaLogo} url='https://chrisstore.co/plugin/85'/>
+        <WelcomeFeature name="N4 Bias Field Correction" img={antLogo} url='https://chrisstore.co/plugin/77'/>
+        <WelcomeFeature name="Civet" img={civetLogo} url='https://chrisstore.co/plugin/2'/>
       </div>
       <div className="welcome-scroll-caret">
-        <Icon name="angle-down" />
+        <AngleDownIcon/>
       </div>
     </div>
     <div className="row no-flex">
