@@ -1,5 +1,6 @@
 import React from 'react';
-import { Button } from 'patternfly-react';
+import Button from '../../../Button';
+
 import './WelcomeChRIS.css';
 import chrisBanner from '../../../../assets/img/chris_logo-ribbon.png';
 
@@ -43,8 +44,7 @@ const WelcomeChRIS = () => (
     <div className="row welcome-chris-btn-row">
       <div>
         <Button
-          bsStyle="default"
-          bsSize="large"
+          variant="secondary"
           className="welcome-chris-btn welcome-chris-more-btn"
           onClick={()=>{window.open('https://www.chrisproject.org')}}
         >
@@ -53,8 +53,7 @@ const WelcomeChRIS = () => (
       </div>
       <div>
         <Button
-          bsStyle="default"
-          bsSize="large"
+          variant="secondary"
           className="welcome-chris-btn"
         >
           Don{"'"}t have access to ChRIS?
